Fix crash when building with g++ 4.3 or newer
This adds a true copy constructor to StdBuf and StdStrBuf.
Unfortunately, this constructor is sometimes prefered to the
move constructor when it wasn't with older compilers, always
referencing the original instead of copying makes the engine
crash, and omitting the copy constructor just creates an
implicit one which also causes crashes.