Fixed a glib warning on startup about the main loop being active already
The exact warning message was

GLib-WARNING **: g_main_context_prepare(): main loop already active in another
thread

The problem was that there is code in C4Console.cpp which iterates the Glib
main loop to update the Console window while loading the game. This was not
synchronized with CGLibProc - we can't iterate the mainloop when the CGLibProc
already does.