win32/gl: Enable switch to fullscreen in debug builds also
Furthermore, I've disabled the window frame while rendering with GL, because
switching from fullscreen to a window with a border and title bar tries to
draw under the border instead of in the client area only. I'm fully aware
this is a hackaround, but if you're using Windows and OpenGL while NOT
testing the mesh branch, it's your own fault.