Clean the windowed mode configuration switch
Instead of doing the resolution switch deep inside the OpenGL initialisation
code, the Application does it on startup, and the code responsible for
runtime resolution switches is responsible for the resolution-dependant
initialisations.

Unfortunately, the D3D code uses two different but similar pieces of code
for both cases, which need to be merged.