Finally move the Log() declaration where it belongs
In the old days, the code was split in two parts - C4* files in one part
and the other files. Long, long ago, the other files were presumably used
by other programs, but that wasn't true since the GWE at least. But the
split was maintained, except for the Log functions. Now that everything
is one giant bunch anyway, the Log functions can be declared in C4Log.h
again.