X11: If there are already events in the queue, signal a timeout
Apparently, software rendering drains the X11 fd. The input code relied on
nothing reading from the fd outside of the event handler, so that the fd
was always readable when new events had arrived. Work around this by
signalling a timeout when there are already events in the internal X11
event queue.