Don't rely on FIXED rounding for movement (#200)
There was an assumption that when rounding a FIXED to an integer A
then adding 1 to the FIXED and rounding the result to an integer B
then B=A+1. This is not true in one special case though: As rounding
is done away from zero -0.5 is rounded to -1 and +0.5 is rounded to +1,
breaking the assumption mentioned above.