Avoid deprecation warnings for static C4ID:: vars
Wrapped the initialisation parameters for the static C4ID member
variables of the C4ID class in std::string( ) temporaries so that the
deprecated const char (&)[] overload of the constructor is not selected.
Since refering to these variables already emits warnings, the warnings
for their definitions are a bit superfluous.