C4Real: Split into multiple implementations
Each implementation gets their own header and class to separate them.
The class is wrapped by C4Real, a template class with the requested
implementation as template argument. This is to ensure a fixed interface
regardless of the underlying implementation. C4Real expects its template
class to mostly have the same interface as a standard C++ float; differences
are handled by template specialization.