C4Real: Implement floating point calculation via SSE
Contrary to x87 FPU floating point calculations, using the SSE unit
should be networking safe, as the internal precision of the SSE unit
is 32 bits when using single precision values. Thus, the compiler's
decision on when to spill registers to memory doesn't influence the
end results of the calculation.

Sine and cosine are calculated using a Taylor approximation after
pulling the input values down to an interval of +-pi/4, with a maximum
relative error of 1.1e-6, or absolute error of 6.0e-8.