win32: Correctly recode incoming wide chars to UTF-8
Surprise: When MSDN says "The WM_CHAR message uses Unicode Transformation
Format (UTF)-16.", it doesn't actually mean that. It means "If your window
is unicode aware, we send UTF-16, otherwise we send default codepage
encoded chars". Except that it doesn't say that anywhere.

So now we recode from the default codepage to UTF-16, and then to UTF-8.