Rewrite ParseCommandLine to work on argc/argv instead of a single string
Using the standard utility function on windows to split the line and the
operating system provided argc/argv on the others.

Also move the function to C4Application. It's not a huge win because the
flags are spread out between C4Game and C4Application pretty evenly, but
eventually most of them should move to a more appropriate place.