CPU scaler case selection smarter in corner cases
Having the calculation in CPU actually allows us to perform more
clever checks. If we have three materials meeting at one point,
we can select one material and make sure that all others pick a
case that's at least consistent to the edge of that material. This
obviously means that the edges between the remaining materials
won't be scaled properly (essentially falling back to GPU scaling),
but that's less disruptive to the result than having the decisions
of neighbouring pixels in disagreement.