MSVC10: Avoid numerous "reload project?" prompts on CMakeLists change
Change the .sln file to make Visual Studio reload that; this also implies
reloading all of the child projects. This way, we don't get a reload prompt for
each loaded project.
On non-Express versions of MSVC, the CMake macros still run. I don't think there
is a way to disable them, but you can copy the macro body of StopBuilds to that
of ReloadProjects to only get a single dialog box.