Some fade tweaking
The slower we fade out, the less quality we get on the fade - which really
shows with normal mapping on flat materials (e.g. a Clonk swimming in Water).
I think a "smooth" factor of 8 is the best compromise here. Maybe I'll find
a better way to do this eventually.

Also reduced quality of landscape approximation. I couldn't see the difference,
so it should be reasonably safe.