Producers: consume required material (#885)
Instead of just checking whether the material is available, also
consume it.

I'm not quite certain whether we should return the removed material
along with the amount removed (to match the return value of the barrel's
GetLiquid), but since the bucket is not currently accepting wildcards
I don't think it's necessary.