Win32: Mark OC "DPI aware" (#891)
Technically this mark is incorrect; we do not handle things differently
on high DPI settings. On the other hand, marking OC DPI aware means that
Windows won't try to scale our drawing, clipping menus and rendering in
the process.

Somewhere in the future, OpenClonk should also add proper DPI awareness
and change the default zoom and menu scaling to match.