Make C4Material.cpp independent of C4Game
This allows compiling and linking with it without pulling in whole
C4Game. This is useful for the mape map editor.

The only place where C4Game was used in C4Material.cpp was in
CrossMapMaterials where the name of the earth material is read from
Game.C4S. The code has been changed such that the earth material name is
now a parameter to the CrossMaterialMap function.