Rename game binary to "openclonk" (#830)
Since the name of the game is OpenClonk, having the binary named
differently is at least confusing. At worst it conflicts with the
trademark license granted by RedWolf Design.

The Mac build will still have to be fixed because the installer
template .dmg file is not editable on non-Macintosh systems.