VolcanoEscapeEXTREME: Close sides
When sides are open and the volcano reaches the central mountain, lava tends to flow out of the map while the volcano continues to grow further up. This looks rather stupid.

The current solution may also look a bit weird (as the volcano now grows over the mountain top), but is preferable both visually and from a gameplay perspective.