custom GUIs: replace position properties X/Y/Wdt/Hgt with Left/Right/Top/Bottom that take strings instead of arrays, also change absolute unit from pixels to em
Before you'd define the X coordinate of the top-left corner of a window with X = [500, -40], now you do Left = "50%-4em". The em values currently do not scale with font-size yet, but it is important to do the formal change as early as possible to not have to do all work twice. Scaling of font and thus menus can be implemented later.