custom guis: fixed some glitches that could happen when nesting layouts and some other minor issues
When the inner layouts adjusted their sizes, the outer layouts would not refresh (glitches occured with GUI_FitChildren). There are some other minor issues fixed as well, like the scroll bar appearing when the children height is exactly the parent's (and the parent uses a layout) or margins not being correctly handled by layouts or Vertical/Grid layout breaking when the child had relative sizes != 0.