Crash Landing: Minor landscape rework. (#1138)
It's the second mission so it's supposed to be easy. But the huge firestone reservoir just beside the bottom granite were just inviting to blast through there.

Also removed some brick areas to make the pumping path less forced. You can now also reroute the lava lake into your starting caves if you want (although this is quite dangerous because you need a lot of space and you're screwed if it fails).